Add REL32 relocs, fix non-zero addend relocs
Add REL32 relocation types as this is emitted by gcc. Note that the
underlying linker in the game does not process REL32 entries properly so
we have to resolve all of these at static link time, otherwise throw an
error.

Relocation resolution does not add non-zero addends in RELA entries properly to
make a relocation entry, so we fix that.